+++ apps/satshoot/src/routes/my-offers/+page.svelte
@@ -7,7 +7,7 @@ import { TabGroup, Tab } from '@skeletonlabs/skeleton';
 import { offerTabStore } from "$lib/stores/tab-store";
 
 import { OfferStatus, type OfferEvent } from '$lib/events/OfferEvent';
-import { allTickets, myOffers } from '$lib/stores/troubleshoot-eventstores';
+import { allTickets, myOffers } from '$lib/stores/freelance-eventstores';
 import OfferCard from '$lib/components/Cards/OfferCard.svelte';
 import { type TicketEvent } from '$lib/events/TicketEvent';
 import SearchIcon from "$lib/components/Icons/SearchIcon.svelte";